tests/wpt/tests/css/css-conditional/container-queries/scroll-state/at-container-scrollable-parsing.html

<!DOCTYPE html>
<title>CSS Conditional Test: @container scrollable query parsing</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#container-rule">
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#scrollable">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-conditional/container-queries/support/cq-testcommon.js"></script>
<div style="container-name:name;container-type:scroll-state">
  <main id="cq-main"></main>
</div>
<script>
  setup(() => assert_implements_scroll_state_container_queries());

  test_cq_condition_known('scroll-state(scrollable)');
  test_cq_condition_known('scroll-state(scrollable: none)');
  test_cq_condition_known('scroll-state(scrollable: top)');
  test_cq_condition_known('scroll-state(scrollable: left)');
  test_cq_condition_known('scroll-state(scrollable: bottom)');
  test_cq_condition_known('scroll-state(scrollable: right)');
  test_cq_condition_known('scroll-state(scrollable: inline-start)');
  test_cq_condition_known('scroll-state(scrollable: inline-end)');
  test_cq_condition_known('scroll-state(scrollable: block-start)');
  test_cq_condition_known('scroll-state(scrollable: block-end)');
  test_cq_condition_known('scroll-state(scrollable: block)');
  test_cq_condition_known('scroll-state(scrollable: x)');
  test_cq_condition_known('scroll-state(scrollable: y)');
  test_cq_condition_known('scroll-state(scrollable: inline)');
  test_cq_condition_known('(scroll-state(scrollable: inline-end))');
  test_cq_condition_known('scroll-state((scrollable: left))');
  test_cq_condition_known('scroll-state(not ((scrollable: bottom) and (scrollable: right)))');
  test_cq_condition_known('scroll-state((scrollable: left) or (scrollable: top))');

  test_cq_condition_unknown('scroll-state(scrollable: auto)');
  test_cq_condition_unknown('scroll-state(scrollable: true)');
  test_cq_condition_unknown('scroll-state(style(scrollable: left))');
  test_cq_condition_unknown('style(scroll-state(scrollable: left))');
  test_cq_condition_unknown('scroll-state(scrollable:)');
</script>
